requirements.txt
aiogram
pydantic-settings
openai == 0.28.0
python-dotenv
gtts
